% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{add_wave}
\alias{add_wave}
\title{Add a wave of data to a panel}
\usage{
add_wave(panel, data, wave, ...)
}
\arguments{
\item{panel}{A panel object}

\item{data}{A \code{data.frame}-like object}

\item{wave}{The wave tag associated with this dataset}

\item{...}{Other arguments passed to methods}
}
\description{
Adds a wave of data to a panel, assigned to wave \code{wave}
}
