% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{add_waves}
\alias{add_waves}
\title{Add multiple waves of data to a panel}
\usage{
add_waves(panel, data, ...)
}
\arguments{
\item{panel}{A panel object}

\item{data}{A list of \code{data.frame}-like objects, with the list names being
wave tags for their respective datasets}

\item{...}{Other arguments passed to methods}
}
\description{
Add multiple waves of data to a panel
}
