% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{amend_wave}
\alias{amend_wave}
\title{Modify a wave's dataset}
\usage{
amend_wave(x, wave_id, wave_db, ...)
}
\arguments{
\item{x}{A panel, currently an unhomogenized panel}

\item{wave_id}{The desired wave tag}

\item{wave_db}{The amended dataset}

\item{...}{Other parameters passed to methods}
}
\description{
Modify a wave's dataset
}
