% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-mapping.R
\name{default_panel_mapping_schema}
\alias{default_panel_mapping_schema}
\title{View the default panel mapping schema}
\usage{
default_panel_mapping_schema()
}
\description{
This schema outlines the column names used by default during panel mapping
creation, when not overridden with the \code{.schema} parameter in
\code{panel_mapping()}.
}
