% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{enpanel}
\alias{enpanel}
\title{Collect multiple waves of data into a panel}
\usage{
enpanel(name, ..., .id_col = "id", .waves_col = "wave")
}
\arguments{
\item{name}{The name of the panel data, e.g. an assessment name}

\item{...}{The waves of the panel. If unnamed arguments are used, the
parameter order will be used as the wave tags. If named parameters are
used, the names of the parameters will be used as the wave tags.}

\item{.id_col}{The name of the column that has the observed participant IDs.
Multiple names can be used if multiple columns uniquely identify
participants.}

\item{.waves_col}{The column name for the wave tags once the panel is
homogenized and bound together into a single \code{data.frame}}
}
\value{
An \code{unhomogenized_panel} object
}
\description{
Collect multiple waves of data into a panel
}
