% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issues}
\alias{issues}
\title{View homogenization issues}
\usage{
issues(x)
}
\arguments{
\item{x}{A panel object}
}
\description{
Some issues may be uncovered over the course of homogenization. Rather than
halting execution immediately upon encountering these problems, these issues
are stored within the panel object. Use this function to view the issues
with a panel.
}
