% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-mapping.R
\name{panel_mapping}
\alias{panel_mapping}
\title{Create a panel mapping}
\usage{
panel_mapping(df, waves, .schema = list())
}
\arguments{
\item{df}{A \code{data.frame} that has variable mapping information}

\item{waves}{A vector of all known waves recorded in the mapping file}

\item{.schema}{A list of column names that outline the structure of the
mapping file. \code{default_panel_mapping_schema()} has the default settings.}
}
\description{
Create a panel mapping
}
