% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelcleaner-package.R
\docType{package}
\name{panelcleaner-package}
\alias{panelcleaner}
\alias{panelcleaner-package}
\title{panelcleaner: An interactive interface to homogenize messy panel data into a long format}
\description{
Panel data structure can sometimes change over the course of data collection, which can be a real nuisance when cleaning data. This package aims to document the state of all waves of data in a comprehensive manner and then homogenize the data into a long dataset ready for data transformations.
}
\details{
The panelcleaner package addresses data structure issues with longitudinal
data by defining two data structures: a \code{tk_panel} class that is a wrapper
around the waves of \emph{tabular} data, and a \code{panel_map} class which is a type
of \code{\link[base:data.frame]{data.frame}} that outlines the raw data content
(variable names and categorical levels) and a common "homogenized" form that
is consistent across all waves.

Resources for workflow:
\itemize{
\item Defining a panel from waves of data: \code{\link[=enpanel]{enpanel()}}, \code{\link[=add_wave]{add_wave()}}
\item Accessing waves of data from an unhomogenized panel: \code{\link[=wave]{wave()}}
\item Defining a panel mapping: \code{\link[=panel_mapping]{panel_mapping()}}
\item Adding a panel mapping to a panel: \code{\link[=add_mapping]{add_mapping()}}
\item Homogenize a panel: \code{\link[=homogenize_panel]{homogenize_panel()}}
\item Combine waves and extract long-form \code{data.frame}: \code{\link[=bind_waves]{bind_waves()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Global-TIES-for-Children/panelcleaner}
  \item Report bugs at \url{https://github.com/Global-TIES-for-Children/panelcleaner/issues}
}

}
\author{
\strong{Maintainer}: Patrick Anker \email{psanker@nyu.edu} (\href{https://orcid.org/0000-0003-2302-0445}{ORCID})

Other contributors:
\itemize{
  \item Global TIES for Children (https://research.steinhardt.nyu.edu/ihdsc/global-ties) [copyright holder]
}

}
\keyword{internal}
