% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{wave}
\alias{wave}
\title{Extract a wave dataset from a panel object}
\usage{
wave(x, wave_id, ...)
}
\arguments{
\item{x}{A panel, currently an unhomogenized panel}

\item{wave_id}{The wave tag to which the dataset is assigned}

\item{...}{Other parameters passed to other methods}
}
\description{
Gets the \code{data.frame} assigned to \code{wave_id} from a panel.
}
