% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenization.R
\name{homogenize_panel}
\alias{homogenize_panel}
\alias{bind_waves}
\title{Homogenize all waves to consistent structure}
\usage{
homogenize_panel(panel, mapping = NULL, allow_issues = FALSE, ...)

bind_waves(panel, allow_issues = FALSE, ...)
}
\arguments{
\item{panel}{An unhomogenized panel}

\item{mapping}{A panel mapping. If NULL, a panel mapping must be attached
to the \code{panel} object using \code{add_mapping()}}

\item{allow_issues}{A logical flag to halt execution if homogenization issues
are encountered. A good safeguard to ensure that your data is consistent.}

\item{...}{Parameters to be used for context, usually for defining a panel
schema}
}
\value{
An \code{unhomogenized_panel} that is ready to be homogenized using
\code{bind_waves()}
}
\description{
Once all waves are collected into a single \code{unhomogenized_panel} object, this
will homogenize variable names and, where applicable, categorical codings
according to a panel mapping.
}
\section{Functions}{
\itemize{
\item \code{bind_waves}: Bind waves into a homogenized panel after
successful homogenization
}}

